bearer-qmi: avoid increasing self reference during dispose()
The reset_bearer_connection() method is called during dispose, and if
a link had been created during the connection attempt, we would be
increasing the reference to the bearer object during dispose(), which
should never happen.

Just avoid passing any callback to the cleanup_link() method, as we
really were using that only for logging purposes.