iface-modem: new optional FCC unlock step in the power state update
There are devices that come locked before they can be put online.
Until now we had a specific implementation for this in the generic QMI
modem, but we should have it in a more generic way for any kind of
modem.

(cherry picked from commit 6ca75c76db277b89325e1888ac445cbfebe33039)