broadband-modem-mbim: skip switch() when selecting MBIM services
Having a switch() for the MBIM services when processing indications
forces us to update it on every new MBIM service added to libmbim,
because we build with -Wswitch-enum by default.

This warning type is extremely useful, and we should not stop using
it, so let's simplify a bit the indication handling code and skip
using a switch().

There are right now only 4 different service indications expected, so
it shouldn't be a big deal.