Revert "foxconn: add new MBIM QDU firmware update method support"
This reverts commit da23f9cd7a7642e8cf8245952a7403bb30fca1b8.

The stable 1.16 branch does not build-depend on any libmbim git master
version, so we should not report the QDU firmware update method as
supported in MM at this point.

It would not have been a big problem if we did expose the QDU firmware
update method, because the logic in fwupd checks the MM and libmbim
versions separately, but I think it's better to skip reporting QDU as
supported ourselves for consistency with the minimum libmbim version
we require.