build,meson: Improve plugins helper library building
To avoid those cases where a plugin does not need a helper library
an empty array is created.

However, when a helper library has to be built, it is not necessary
to include the created static library in an array.

(cherry picked from commit 80f274b37159777a462d7584fb4fc82876229ac2)