build,meson: fix building shared utils and plugins
When building shared utils modules or plugins, we should not link with
the static libraries built under src/, because that would mean we're
including their symbols as exported by the modules, and it'll break in
runtime.

Instead, just define the correct set of 'include_directories' to be
used in each plugin.

Fixes https://gitlab.freedesktop.org/mobile-broadband/ModemManager/-/issues/436