ci: bump image template to refresh CA certificates
At the end of September 2021, Let's Encrypt's DST Root CA X3
certificate expired, and the new ISRG Root X1 took over.
Update the CI template so that the new certificate is available and we
can keep on cloning the libmbim/libqmi/libqrtr-glib repos without
failures.

We don't refresh the date tag in this template bump, so that we don't
confuse with the date tag bumps done in the image used for git master
and MM 1.18, which are different due to the meson build.